Add Framework target
This adds a framework target, which will allow this project to be built
for use with Git Submodules as well as with Carthage.

- Extracted LUKeychainAccessAccessibility enum to its own file to sever
  a circular import between LUKeychainAccess and LUKeychainServices
- Removed the library target used in testing in favor of a framework
  target. This target is iOS 8+ only, but it hasn't changed the
  deployment target for CocoaPods installation.
- Shared the LUKeychainAccess scheme
