+++ LUKeychainAccess/LUKeychainServices.h
@@ -1,7 +1,7 @@
 // A wrapper for Keychain Services using the Facade pattern: http://en.wikipedia.org/wiki/Facade_pattern
 
 #import <Foundation/Foundation.h>
-#import "LUKeychainAccess.h"
+#import "LUKeychainAccessAccessibility.h"
 
 @interface LUKeychainServices : NSObject
