+++ Unit-Tests/LUKeychainAccessSpec.m
@@ -1,10 +1,9 @@
+@import LUKeychainAccess;
 #import "Kiwi.h"
-#import "LUKeychainAccess.h"
-#import "LUKeychainServices.h"
 #import "LUTestErrorHandler.h"
 
 SPEC_BEGIN(LUKeychainAccessSpec)
-
+  
 describe(@"LUKeychainAccess", ^{
   __block LUKeychainAccess *keychainAccess;
   __block LUKeychainServices *keychainServices;
